# metal: template for mat-vec multiplication kernels