This commit provides one template for all mat-vec kernels, making it easier for us to implement multiplication kernels between two quantized types in future. 
Also, speedup `Q_K` and `F16` inference speed a little bit.

| model                           | backend   |    threads | test       |       master       t/s | PR       t/s | diff
| ------------------------------ | ---------| ---------: | ---------- | ---------------: |---------------: |---------------: |
| llama 30B Q4_0 |  Metal      |        4 | pp 512     |     68.46 ± 0.02 | 67.85 ± 0.07 | -0.9%
| llama 30B Q2_K |  Metal      |        4 | pp 512     |     63.85 ± 0.01 |64.20 ± 0.01 | 0.5%
| llama 30B Q3_K_S | Metal      |       4 | pp 512     |     61.02 ± 0.01 |60.53 ± 0.01 |-0.8%
| llama 30B Q4_K_S |  Metal      |        4 | pp 512     |     61.58 ± 0.03 |  62.48 ± 0.01 |1.4%
| LLaMA2 34B Q5_K_S  | Metal      |          4 | pp 512     |     45.97 ± 0.04 |47.92 ± 0.01 |4.2%
| LLaMA2 34B Q6_K |  Metal      |         4 | pp 512     |     43.52 ± 0.03 | 47.74 ± 0.00 |9.7%
| llama 30B Q4_0 | Metal      |         4 | tg 128     |     16.10 ± 0.01 | 16.06 ± 0.01 | -0.2%
|llama 30B Q2_K  |Metal      |        4 | tg 128     |     15.41 ± 0.00 |17.46 ± 0.00 |13.3%
| llama 30B Q3_K_S| Metal      |          4 | tg 128     |     11.36 ± 0.00 |   13.17 ± 0.00 |16%
| llama 30B Q4_K_S  |  Metal      |          4 | tg 128     |     14.48 ± 0.01 |  15.53 ± 0.01 |7.3%
| LLaMA2 34B Q5_K_S  |  Metal      |          4 | tg 128     |      9.87 ± 0.00 |  11.89 ± 0.01 |20.4%
| LLaMA2 34B Q6_K | Metal      |           4 | tg 128     |      7.95 ± 0.00 |   10.91 ± 0.01 |37.2%
| falcon 7B F16 | Metal      |         4 | pp 512     |    154.59 ± 0.53 | 154.69 ± 0.31 |0.0%
| falcon 7B F16 | Metal      |        4 | tg 128     |     17.45 ± 0.01 |  22.77 ± 0.02 |30.5%

`QK_K=256` was tested on llama models, and `QK_K=64` was tested on openllama models. People are welcome to test this pull request.